♻️ Refactor: Replace Emoji components with Tag and simplify emoji picker
